Profile Part 2

+++ app/src/main/kotlin/io/github/wykopmobilny/api/entries/EntriesApi.kt
@@ -17,26 +17,26 @@ interface EntriesApi {
     val entryVoteSubject: PublishSubject<EntryVotePublishModel>
     val entryUnVoteSubject: PublishSubject<EntryVotePublishModel>
 
-    fun voteEntry(entryId: Int, notifySubject: Boolean = true): Single<VoteResponse>
-    fun unvoteEntry(entryId: Int, notifySubject: Boolean = true): Single<VoteResponse>
-    fun voteComment(commentId: Int): Single<VoteResponse>
-    fun unvoteComment(commentId: Int): Single<VoteResponse>
+    fun voteEntry(entryId: Long, notifySubject: Boolean = true): Single<VoteResponse>
+    fun unvoteEntry(entryId: Long, notifySubject: Boolean = true): Single<VoteResponse>
+    fun voteComment(commentId: Long): Single<VoteResponse>
+    fun unvoteComment(commentId: Long): Single<VoteResponse>
     fun addEntry(body: String, wykopImageFile: WykopImageFile, plus18: Boolean): Single<EntryResponse>
     fun addEntry(body: String, embed: String?, plus18: Boolean): Single<EntryResponse>
-    fun addEntryComment(body: String, entryId: Int, embed: String?, plus18: Boolean): Single<EntryCommentResponse>
-    fun addEntryComment(body: String, entryId: Int, wykopImageFile: WykopImageFile, plus18: Boolean): Single<EntryCommentResponse>
-    fun markFavorite(entryId: Int): Single<FavoriteResponse>
-    fun deleteEntry(entryId: Int): Single<EntryResponse>
-    fun editEntry(body: String, entryId: Int): Single<EntryResponse>
-    fun editEntryComment(body: String, commentId: Int): Single<EntryCommentResponse>
-    fun deleteEntryComment(commentId: Int): Single<EntryCommentResponse>
-    fun voteSurvey(entryId: Int, answerId: Int): Single<Survey>
+    fun addEntryComment(body: String, entryId: Long, embed: String?, plus18: Boolean): Single<EntryCommentResponse>
+    fun addEntryComment(body: String, entryId: Long, wykopImageFile: WykopImageFile, plus18: Boolean): Single<EntryCommentResponse>
+    fun markFavorite(entryId: Long): Single<FavoriteResponse>
+    fun deleteEntry(entryId: Long): Single<EntryResponse>
+    fun editEntry(body: String, entryId: Long): Single<EntryResponse>
+    fun editEntryComment(body: String, commentId: Long): Single<EntryCommentResponse>
+    fun deleteEntryComment(commentId: Long): Single<EntryCommentResponse>
+    fun voteSurvey(entryId: Long, answerId: Int): Single<Survey>
 
     fun getHot(page: Int, period: String): Single<List<Entry>>
     fun getStream(page: Int): Single<List<Entry>>
     fun getActive(page: Int): Single<List<Entry>>
     fun getObserved(page: Int): Single<List<Entry>>
-    fun getEntry(id: Int): Single<Entry>
-    fun getEntryVoters(id: Int): Single<List<Voter>>
-    fun getEntryCommentVoters(id: Int): Single<List<Voter>>
+    fun getEntry(id: Long): Single<Entry>
+    fun getEntryVoters(id: Long): Single<List<Voter>>
+    fun getEntryCommentVoters(id: Long): Single<List<Voter>>
 }
