+++ app/src/main/kotlin/io/github/wykopmobilny/ui/modules/input/entry/edit/EditEntryView.kt
@@ -3,5 +3,5 @@ package io.github.wykopmobilny.ui.modules.input.entry.edit
 import io.github.wykopmobilny.ui.modules.input.BaseInputView
 
 interface EditEntryView : BaseInputView {
-    val entryId: Int
+    val entryId: Long
 }
